Added datasources

+++ src/main/java/dev/autonu/framework/common/model/ClientUserAssociation.java
@@ -3,5 +3,5 @@ package dev.autonu.framework.common.model;
 /**
  * @author autonu2X
  */
-public record ClientUserAssociation(Long clientId, Long userId, String username) {
+public record ClientUserAssociation(Integer clientId, Long userId, String username) {
 }
